% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_atp_data.R
\name{build_atp_data}
\alias{build_atp_data}
\title{Build ATP dataset using Jeff Sackman Github data}
\usage{
build_atp_data(path)
}
\arguments{
\item{path}{file path for directory of yearly match data csvs}
}
\value{
dataframe of yearly ATP match data with season id
}
\description{
Build ATP dataset using Jeff Sackman Github data
}
